Extract `new` into own module
